fixlets and improvments after little trip to old solaris
 .mapfile
	bring symbols in the mapfile up to date
 G2Acceptor.c
	make symbol static
	fix ipv4 vs. ipv6 handling
	use setsocketopt option only if avail
	pass the right structure length to function, or libc may barf (EINVAL)
	is* casting to int
 G2Connection.c
	is* casting to int
 lib/vsnprintf.c
	is* casting to int
	make hexstrings unique
 G2Handler.c
	remove unused header
	make symbol static
 G2KHL.c
	missed header
	port fallback
	print resolver errors
	is* casting to int
 G2Packet.c
	double const
 G2UDP.c
	signature cast
	use setsocketopt option only if avail
	pass the right structure length to function, or libc may barf (EINVAL)
 lib/module.make
 Makefile
	incorporate bin2o options
	split and added/fixed options
	more -D for host programs
	add missed file to tar (ouch!)
 arflock.c
	warning fix
 bin2o.c
	fix sun as
 config.h
	more options
 lib/aes_tab_gen.c
 lib/five_tab_gen.c
	dirty include fix
 lib/arm/atomic.h
 lib/sparc64/atomic.h
	ahhh, fix the typo!
 lib/bitfield_rle.c
	we have a func for that, don't use the GCC >= 4 call
 lib/combo_addr.h
	solrais does not define an accessor for the 32 bit union
 lib/generic/cpy_rest.c
	*facepalm* keep i intact to to pointer arith.
 lib/my_epoll_devpoll.c
 lib/generic/memneg.c
 lib/generic/memxorcpy.c
	fix syntax error
	fix errors
 lib/generic/strnpcpy.c
	use the right pointer...
 lib/hthash.h
	remove here, belongs into other.h
 lib/hzp.c
	fix include order
 lib/inet_ntop.c
	solaris has no 16 bit fields in the union, don't use them
 lib/other.h
	more porting foo, extend endian handling
 lib/udpfromto.c
	errors, porting foo

  lib/generic/mempopcnt.c
	provide a genric new & improved, corrected base func, where archs can
	use it
 lib/alpha/mempopcnt.c
 lib/ia64/mempopcnt.c
 lib/sparc64/mempopcnt.c
	use generic func, only deliver pop impl.
 lib/ppc/mempopcnt.c
	use generic func, only deliver pop impl.
	*headdesk* vec_perm, the poor mans lookup table...
 lib/x86/mempopcnt.c
	woooo!
	use generic func for x86
	fix mmx and SSE2 impl.
	remove computed jump from SSE2
	add SSSE3 impl., PSHUFB can be used for a lookup table
	write main loop of SSE4a impl., the compiler got crazy
	 (unroll 8 times, then spill every result to stack, huh??)

 lib/ppc/popcountst.c
	let the compiler mull and shift, he knows he does

+++ arflock.c
@@ -1,7 +1,7 @@
 /* arflock.c
  * ar proxy to lock the .a so concurrent access is possible
- * 
- * Copyright (c) 2006, Jan Seiffert
+ *
+ * Copyright (c) 2006-2009 Jan Seiffert
  *
  * This file is part of g2cd.
  *
@@ -9,17 +9,17 @@
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation; either version 2
  * of the License, or any later version.
- * 
+ *
  * g2cd is distributed in the hope that it will be useful, but
  * WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with g2cd; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston,
  * MA  02111-1307  USA
- * 
+ *
  * $Id: $
  */
 
@@ -65,7 +65,7 @@ int main(int argc, char **argv)
 		perror(__FILE__ ", " mkstr(__LINE__) ", open() gate" );
 		exit(EXIT_FAILURE);
 	}
-	ret_val = sprintf(pid_buf, "%i\n", getpid());
+	ret_val = sprintf(pid_buf, "%i\n", (int)getpid());
 	if(ret_val != write(fd_gate, pid_buf, ret_val)) {
 		/* nothing we can do about, it's only nice to put our pid inside */
 	}
